requirements.txt
click==8.1.7
pandas==2.1.1
pyyaml==6.0.1
requests==2.31.0
